requirements.txt
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
openai==0.27.8
werkzeug==2.0.1
scikit-learn==1.0.2
numpy==1.21.2
joblib==1.1.0